third_party/WebKit/LayoutTests/fast/dom/idl-callback-function-unittest.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>
var callbackFunctionTest = internals.callbackFunctionTest();

test(function() {
    var callback1 = function(msg1, msg2) {
        return msg1 + ', ' + msg2;
    };
    assert_equals(callbackFunctionTest.testCallback(callback1, 'hello', 'world'), 'SUCCESS: hello, world');

    var callback2 = function(msg1, msg2) {
        return msg1 + msg1 + ' ' + msg2 + msg2;
    };
    assert_equals(callbackFunctionTest.testCallback(callback2, 'hello', 'world'), 'SUCCESS: hellohello worldworld');
}, 'Callback function which takes two strings');

test(function() {
    var divElement = document.createElement('div');
    var addInnerHTML = function(d) {
        d.innerHTML = 'hello';
    };
    callbackFunctionTest.testInterfaceCallback(addInnerHTML, divElement);
    assert_equals(divElement.innerHTML, 'hello');
}, 'Callback function which takes a interface');

test(function() {
    var callback = function() {
        assert_equals(this, callbackFunctionTest);
    };
    callbackFunctionTest.testReceiverObjectCallback(callback);
}, 'Callback function which takes receiver object');
</script>
